requirements.txt
langchain
faiss-cpu
PyPDF2
python-dotenv
streamlit
langchain-google-genai
